Require curly braces for all blocks (#105)
# Justification

Fixes #86.

# Implementation

Enable the
[brace-style](https://eslint.org/docs/latest/rules/brace-style) and
[curly](https://eslint.org/docs/rules/curly) ESLint rules to match the
decision described in the issue linked above. Now braces need to be
present and the body must be on a new line:
```ts
// allowed
if (foo) {
    doSomething();
}

// disallowed
if (foo) { doSomething(); }
if (foo) doSomething();
if (foo) 
    doSomething();
```

# Testing

Manually modified test files within the repo and verified that lint
errors are reported for both JS and TS code if you have an `if` without
braces.