Revert Angular 15 support (#132)
My previous PR (#129) did not simplify the Angular 15 upgrade as I had
hoped it would. Instead, it caused more problems because npm is having
trouble resolving some of the peer dependencies that need to be
installed. Therefore, I am reverting the PR that added Angular 15
support, and I will make a separate PR after this one is merged that
changes the Angular support of this package to _only_ Angular 15.

Resolves #131 

I've tested this change in an Angular 14 application, and I verified
that the updated package can install without requiring any additional
dependencies to be configured.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>