Allow pascal-case imports and disable consistent-generic-constructors (#162)
Found two rules causing new, unwanted violations in the Nimble repo:
- `@typescript-eslint/consistent-generic-constructors` results in code
like this:

    ```
    private readonly hiddenOptions: Set<ListOption> = new Set();
    ```
reporting violations like "The generic type arguments should be
specified as part of the constructor type arguments". I had tried to
disable this in the prior PR, but was doing so from the wrong config.
- `@typescript-eslint/naming-convention` now covers imported namespaces
(e.g. `import * as Foo from '../foo'`), and it defaults to allowing both
camel and pascal case. However, because we are configuring the
`'default'` selector (which matches anything not otherwise specified) to
only allow camel case, it was disallowing pascal case for those imported
namespaces.