Add Spright chat attribute to Angular i18n ignore list (#165)
## Problem

We keep a list of Nimble attributes that shouldn't trigger the i18n lint
rule. The new Spright chat components didn't appear in this list so once
we create Angular wrappers they would incorrectly trigger that rule.

## Implementation

Add a new section for Spright components and add the `message-type`
attribute to that section.