Update the README sections on flat configuration and Angular. (#173)
I updated an Angular application following the README and have updated
the README for the style guide with important information.

The tests for Angular also need to be expanded to validate that our
configuration and how we document its use works with Angular workspaces
and the VSCode extensions. This will be done in a separate PR.

- Update the README sections on flat configuration and Angular.
- Fix the `tsconfigRootDir` configuration in the tests.

Testing
- Verified that the linter passes
- Verified that the automated tests pass
- Manually verified that the installation instructions in the README
work correctly
- Manually verified that the shared configurations work correctly